requirements.txt
transformers>=4.20.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
datasets>=2.0.0
nltk
scikit-learn
pandas
numpy
tqdm
wandb
